Pin openshift-golang-builder to 1.23 rhel8 tag
In 4f3ee0f, 1.23 was introduced, but just referenced by manifest
digest. Unfortunately the Konflux bot appears to have assumed that it
was the `latest` tag, and saw a "new" (read: several years old) digest
for that tag, and bumped to it in 3798127.

This change specifies the rhel_8_golang_1.23 tag, so that it will only
bump to newer versions of _that_ tag.

This change also changes the registry domain to be
brew.registry.redhat.io, since 68487ea (#194) appears to indicate that
that's preferred/required for Konflux. One can use the `--build-arg`
CLI option to use the registry-proxy locally as needed.